Added cancel button section. Added support for scrubbing between buttons.
